+++ src/main/java/com/yoyomo/domain/user/application/mapper/ManagerMapper.java
@@ -1,6 +1,6 @@
 package com.yoyomo.domain.user.application.mapper;
 
-import com.yoyomo.domain.user.application.dto.ManagerDTO;
+import com.yoyomo.domain.user.application.dto.response.ManagerResponseDTO;
 import com.yoyomo.domain.user.domain.entity.Manager;
 import com.yoyomo.global.config.jwt.presentation.JwtResponse;
 import com.yoyomo.global.config.kakao.dto.KakaoAccount;
@@ -11,7 +11,7 @@ import org.mapstruct.*;
         nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
 public interface ManagerMapper {
 
-    ManagerDTO.Response to(Manager manager, JwtResponse token);
+    ManagerResponseDTO.Response to(Manager manager, JwtResponse token);
 
     @Mapping(target = "name", source = "dto.profile.nickname")
     Manager from(KakaoAccount dto);
